Add Tosca short names and restructure file tree
