--- lidy-js/src/listener/prog.js
@@ -1,5 +1,0 @@
-function exit_main(parsed_rule) {
-    console.log("DEBUG: Enter Main");
-}
-
-export default { exit_main }
