+++ lidy-js/src/tosca/index.js
@@ -1,0 +1,3 @@
+import { parse } from './tosca/parser/parse.js'
+
+export default parse;
