+++ lidy-js/src/tosca/listener/imports.js
@@ -1,4 +1,4 @@
-import { newToscaImport } from '../tosca/imports.js'
+import { newToscaImport } from '../model/imports.js'
 
 function exit_imports(parsed_rule) {
     parsed_rule.value.forEach(val => {
@@ -17,7 +17,6 @@ function exit_imports(parsed_rule) {
             namespace_uri}, 
             val)
         parsed_rule.ctx.prog.current_service_template.imports.push(newImport)
-        // parsed_rule.ctx.prog.imports.push(newImport)
     })
 }
 export default { exit_imports }
