+++ lidy-js/src/tosca/listener/prog.js
@@ -1,0 +1,3 @@
+function exit_main(parsed_rule) {}
+
+export default { exit_main }
