+++ lidy-js/src/tosca/model/namespace.js
