+++ lidy-js/src/tosca/model/prog.js
@@ -1,5 +1,3 @@
-// import namespace from "../listener/namespace"
-
 export class ToscaProg {
     constructor() {
         this.errors = []
@@ -53,42 +51,31 @@ export class ToscaType extends ToscaNode {
         let namespace_name = current_st.namespace.value
 
         // dans le current_st
-        // current_st[getClassnalme()][current_st.namespace + "/" + name]
         if (current_st[category][namespace_name + "/" + name]) {
             parsed_rule.ctx.grammarError('Type collision : ' + this.import_id)
             console.log("Erreur de collision de type");
         } else {
             current_st[category][namespace_name + "/" + name] = this
-            current_st[category][name] = this            
         }
 
         // dans le parent_st
         // si namspace_uri alors namespace = namespace_uri
         // sinon namespace = namespace
         let namespace = (current_st.ns_uri) ? current_st.ns_uri : namespace_name
-        // parents_st[getClassname()][namespace + "/" + name]
         if (parent_st) {
             if (parent_st[category][namespace + "/" + name]) {
                 parsed_rule.ctx.grammarError('Type collision : ');
                 console.log("Erreur de collision de type");
             } else {
                 parent_st[category][namespace + "/" + name] = this
-                if (!parent_st[category][name]) {
-                    parent_st[category][name] = this
-                }
-            }
 
-            // si ns_prefix alors 
-            // parents_st[getClassname()][prefix + "." + name]
+            }
             if (current_st.ns_prefix) {
                 if (parent_st[category][current_st.ns_prefix + "." + name]) {
                     parsed_rule.ctx.grammarError('Type collision : ')
                     console.log("Erreur de collision de type");
                 } else {
                     parent_st[category][current_st.ns_prefix + "." + name] = this
-                    if (parent_st[category][name]) {
-                        parent_st[category][name] = this
-                    }
                 }
             }
         }
@@ -105,9 +92,7 @@ export class ToscaType extends ToscaNode {
         return str;
     }
     static isValid(input, source) {
-        if (typeof (input.derived_from) != 'string' // || 
-        ) {
-
+        if (typeof (input.derived_from) != 'string') { // TODO
             return false
         }
         return true
