+++ lidy-js/src/tosca/tests/data/ToscaExampleSimple.yml
