+++ lidy-js/src/tosca/tests/data/ToscaExampleSimple2.yml
