+++ lidy-js/src/tosca/tests/data/testNomsCourts/B.yml
@@ -1,0 +1,27 @@
+tosca_definitions_version: tosca_simple_yaml_1_3
+
+description: Template for testing imports. 
+
+namespace: https://B.orness.com
+
+metadata:
+  creation_date: 2015-04-14
+  date_updated: 2015-05-01
+  status: developmental
+  template_version: 2.0.17
+  template_author: Jean-Luc Duvivier
+  template_name: My service template
+
+############################
+##### TOSCA NODE_TYPES #####
+############################
+node_types:
+  
+  toto.titi.tata: 
+    derived_from: tosca.nodes.Root
+
+  tutu.plouf: 
+    derived_from: tosca.nodes.Root
+
+  tagada.pouetpouet: 
+    derived_from: tosca.nodes.Root
